Fix type recursion error on Clang
Art's ServiceCacheEntry uses the ServiceCache class, which itself
uses ServiceCacheEntry. This results in an incomplete type error on
Clang. This is "correct" in that the Standard makes no requirement
on behaviour here (undefined). GCC allows this, but it is an
implementation detail of that compiler.

Provide a temporary fix by replacing use of std::map for ServiceCache
by Boost's boost::collection::map type which allows type recursion.